service/llm_chat_models/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
httpx==0.27.0
pydantic==2.6.1
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-pymongo
pymongo
requests
websockets
streamlit==1.31.0
watchdog
psycopg2-binary

